refactor(stop_filter): rename file name
