+++ localization/autoware_stop_filter/src/stop_filter_node.cpp
@@ -1,4 +1,4 @@
-// Copyright 2021 TierIV
+// Copyright 2025 TierIV
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "stop_filter.hpp"
+#include "autoware/stop_filter/stop_filter_node.hpp"
 
 #include <rclcpp/logging.hpp>
